+++ .types.lua
@@ -1,0 +1,13 @@
+---@diagnostic disable: duplicate-set-field
+---@meta
+
+---
+-- This class contains EmmyLua annotations to help
+-- IDEs work with some external classes and types
+---
+
+-- WoW methods
+
+---@class C_TradeSkillUI
+---@field GetRecipeItemLink fun(recipeID: number): string
+C_TradeSkillUI = {}