+++ CHANGES.md
@@ -1,1 +1,5 @@
-- Fix breaking crafts without quality
+- Update ToC version for patch 10.0.5
+- Add chat command to set recraft item by item link
+- Fix experiment checkbox placement in order UI
+- Fix not clearing optional reagent slots in order UI
+- Fix clearing optional slots if crafting UI hasn't been opened yet